Move init to its own service
