+++ packages/cli/lib/cli.ts
@@ -8,7 +8,7 @@ import * as dotenv from 'dotenv';
 
 import { getNangoRootPath, printDebug } from './utils.js';
 import { loadYamlAndGenerate } from './services/model.service.js';
-import { NANGO_INTEGRATIONS_NAME, TYPES_FILE_NAME, exampleSyncName } from './constants.js';
+import { TYPES_FILE_NAME, exampleSyncName } from './constants.js';
 import { compileAllFiles, compileSingleFile, getFileToCompile } from './services/compile.service.js';
 import { getLayoutMode } from './utils/layoutMode.js';
 import { getProviderConfigurationFromPath, nangoConfigFile } from '@nangohq/nango-yaml';
@@ -127,81 +127,6 @@ export function generate({ fullPath, debug = false }: { fullPath: string; debug?
             }
         }
     }
-}
-
-/**
- * Init
- * If we're not currently in the nango-integrations directory create one
- * and create an example nango.yaml file
- */
-export function init({ absolutePath, debug = false }: { absolutePath: string; debug?: boolean }) {
-    const yamlData = fs.readFileSync(path.resolve(__dirname, `./templates/${nangoConfigFile}`), 'utf8');
-    const envData = fs.readFileSync(path.resolve(__dirname, './templates/env'), 'utf8');
-
-    // if currently in the nango-integrations directory then don't create another one
-    const currentDirectory = path.basename(absolutePath);
-    let fullPath: string;
-
-    if (currentDirectory === NANGO_INTEGRATIONS_NAME) {
-        if (debug) {
-            printDebug(`Currently in the ${NANGO_INTEGRATIONS_NAME} directory so the directory will not be created`);
-        }
-        fullPath = absolutePath;
-    } else {
-        fullPath = path.resolve(absolutePath, NANGO_INTEGRATIONS_NAME);
-    }
-
-    if (fs.existsSync(fullPath)) {
-        console.log(chalk.red(`The ${NANGO_INTEGRATIONS_NAME} directory already exists. You should run commands from within this directory`));
-    } else {
-        if (debug) {
-            printDebug(`Creating the nango integrations directory at ${absolutePath}`);
-        }
-        fs.mkdirSync(fullPath);
-    }
-
-    const configFileLocation = path.resolve(fullPath, nangoConfigFile);
-    if (!fs.existsSync(configFileLocation)) {
-        if (debug) {
-            printDebug(`Creating the ${nangoConfigFile} file at ${configFileLocation}`);
-        }
-        fs.writeFileSync(configFileLocation, yamlData);
-    } else {
-        if (debug) {
-            printDebug(`Nango config file already exists at ${configFileLocation} so not creating a new one`);
-        }
-    }
-
-    const envFileLocation = path.resolve(fullPath, '.env');
-    if (!fs.existsSync(envFileLocation)) {
-        if (debug) {
-            printDebug(`Creating the .env file at ${envFileLocation}`);
-        }
-        fs.writeFileSync(envFileLocation, envData);
-    } else {
-        if (debug) {
-            printDebug(`.env file already exists at ${envFileLocation} so not creating a new one`);
-        }
-    }
-
-    const gitIgnoreFileLocation = path.resolve(fullPath, '.gitignore');
-    if (!fs.existsSync(gitIgnoreFileLocation)) {
-        if (debug) {
-            printDebug(`Creating the .gitignore file at ${gitIgnoreFileLocation}`);
-        }
-        fs.writeFileSync(
-            gitIgnoreFileLocation,
-            `dist
-.env
-`
-        );
-    } else {
-        if (debug) {
-            printDebug(`.gitignore file already exists at ${gitIgnoreFileLocation} so not creating a new one`);
-        }
-    }
-
-    generate({ debug, fullPath });
 }
 
 export function tscWatch({ fullPath, debug = false }: { fullPath: string; debug?: boolean }) {
