+++ packages/cli/lib/sync.unit.cli-test.ts
@@ -4,8 +4,9 @@ import os from 'node:os';
 import fs from 'node:fs';
 import yaml from 'js-yaml';
 import stripAnsi from 'strip-ansi';
-import { init, generate } from './cli.js';
+import { generate } from './cli.js';
 import { exampleSyncName } from './constants.js';
+import { init } from './services/init.service.js';
 import { compileAllFiles, compileSingleFile, getFileToCompile } from './services/compile.service.js';
 import parserService from './services/parser.service.js';
 import { copyDirectoryAndContents, removeVersion, fixturesPath, getTestDirectory } from './tests/helpers.js';
